requirements.txt
Django>=3.2,<3.3
wagtail>=2.15,<2.16
django-environ
mysqlclient
